Extract shared BSD hardware classes (NetBSD/OpenBSD dedup)
Merge identical NetBSD/OpenBSD implementations into shared classes
in the unix hardware package:
- BsdUsbDevice (usbdevs -v parsing)
- BsdGraphicsCard (pcidump -v parsing)
- BsdSoundCard (dmesg audio/pci parsing)
- BsdSensors (systat sensor queries)

Eliminates ~280 duplicated lines across 8 removed files.
