+++ oshi-common/src/main/java/oshi/hardware/common/platform/unix/BsdSensors.java
@@ -2,7 +2,7 @@
  * Copyright 2021-2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
-package oshi.hardware.common.platform.unix.openbsd;
+package oshi.hardware.common.platform.unix;
 
 import static oshi.util.Memoizer.defaultExpiration;
 import static oshi.util.Memoizer.memoize;
@@ -15,10 +15,11 @@ import oshi.hardware.common.AbstractSensors;
 import oshi.util.tuples.Triplet;
 
 /**
- * Sensors
+ * Shared sensors implementation for BSDs that use {@code systat} for temperature, fan speed, and voltage (NetBSD,
+ * OpenBSD).
  */
 @ThreadSafe
-public final class OpenBsdSensors extends AbstractSensors {
+public final class BsdSensors extends AbstractSensors {
 
     private final Supplier<Triplet<Double, int[], Double>> tempFanVolts = memoize(Systat::querySensors,
             defaultExpiration());
