--- oshi-common/src/main/java/oshi/hardware/common/platform/unix/netbsd/NetBsdSensors.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2021-2026 The OSHI Project Contributors
- * SPDX-License-Identifier: MIT
- */
-package oshi.hardware.common.platform.unix.netbsd;
-
-import static oshi.util.Memoizer.defaultExpiration;
-import static oshi.util.Memoizer.memoize;
-
-import java.util.function.Supplier;
-
-import oshi.annotation.concurrent.ThreadSafe;
-import oshi.driver.common.unix.bsd.Systat;
-import oshi.hardware.common.AbstractSensors;
-import oshi.util.tuples.Triplet;
-
-/**
- * Sensors
- */
-@ThreadSafe
-final class NetBsdSensors extends AbstractSensors {
-
-    private final Supplier<Triplet<Double, int[], Double>> tempFanVolts = memoize(Systat::querySensors,
-            defaultExpiration());
-
-    @Override
-    public double queryCpuTemperature() {
-        return tempFanVolts.get().getA();
-    }
-
-    @Override
-    public int[] queryFanSpeeds() {
-        return tempFanVolts.get().getB();
-    }
-
-    @Override
-    public double queryCpuVoltage() {
-        return tempFanVolts.get().getC();
-    }
-}
