+++ oshi-core-ffm/src/main/java/oshi/hardware/platform/unix/openbsd/OpenBsdHardwareAbstractionLayerFFM.java
@@ -20,13 +20,13 @@ import oshi.hardware.Sensors;
 import oshi.hardware.SoundCard;
 import oshi.hardware.UsbDevice;
 import oshi.hardware.common.AbstractHardwareAbstractionLayer;
+import oshi.hardware.common.platform.unix.BsdGraphicsCard;
 import oshi.hardware.common.platform.unix.BsdNetworkIF;
 import oshi.hardware.common.platform.unix.BsdPowerSource;
+import oshi.hardware.common.platform.unix.BsdSensors;
+import oshi.hardware.common.platform.unix.BsdSoundCard;
+import oshi.hardware.common.platform.unix.BsdUsbDevice;
 import oshi.hardware.common.platform.unix.UnixDisplay;
-import oshi.hardware.common.platform.unix.openbsd.OpenBsdGraphicsCard;
-import oshi.hardware.common.platform.unix.openbsd.OpenBsdSensors;
-import oshi.hardware.common.platform.unix.openbsd.OpenBsdSoundCard;
-import oshi.hardware.common.platform.unix.openbsd.OpenBsdUsbDevice;
 import oshi.hardware.platform.unix.CupsPrinterFFM;
 
 @ThreadSafe
@@ -49,7 +49,7 @@ public final class OpenBsdHardwareAbstractionLayerFFM extends AbstractHardwareAb
 
     @Override
     public Sensors createSensors() {
-        return new OpenBsdSensors();
+        return new BsdSensors();
     }
 
     @Override
@@ -74,17 +74,17 @@ public final class OpenBsdHardwareAbstractionLayerFFM extends AbstractHardwareAb
 
     @Override
     public List<UsbDevice> getUsbDevices(boolean tree) {
-        return OpenBsdUsbDevice.getUsbDevices(tree);
+        return BsdUsbDevice.getUsbDevices(tree);
     }
 
     @Override
     public List<SoundCard> getSoundCards() {
-        return OpenBsdSoundCard.getSoundCards();
+        return BsdSoundCard.getSoundCards();
     }
 
     @Override
     public List<GraphicsCard> getGraphicsCards() {
-        return OpenBsdGraphicsCard.getGraphicsCards();
+        return BsdGraphicsCard.getGraphicsCards();
     }
 
     @Override
